test: custom drag overlay
